Slash 'n burn
- Remove most of `decode.go` & tests.
  - All interfaces (`Decoder`, `SimpleDecoder`, `CSVReader`), their implementations, and helpers are gone.
  - Nuke `readTo`, `readToWithErrorHandler`, `readEach`, `readEachWithoutHeaders`, and `readToWithoutHeaders`.  They’re overly complicated, duplicative, and unnecessary.
- Remove most `csv.go`.
  - Remove all the global configuration: `selfCSVWriter`, `selfCSVReader`, and their getters/setters.
  - Remove literal dozens of `MarshalBlizzardOfTypesAndOptions` / `UnmarshalBlizzardOfTypesAndOptions` helper functions.
  - Remove `ErrorHandler` type alias, it was only used by the deleted `decoder` code.
- Remove `safe_csv.go`.
  - All it did is wrap `csv.Writer` with a mutex, so you can have concurrent processes writing to the same output, which is a bad idea to start.
  - Just pass your computed values to a writer goroutine with a channel, that’s what they’re for.
  - Or add your own locks around the writer.
  - But this is a bad pattern, if you use it, you should feel bad & deal with the consequences.
  - It doesn’t belong in the library.
- Move `UnmarshalCSVWithFieldsSample`'s `UnmarshalCSVWithFields()` function into the same file as the struct is defined in, rather than the random-ass one it was inexplicably in.

- Remove `safe_csv.go`.  All it does is lock/unlock around
