+++ decode.go
@@ -1,57 +1,14 @@
 package gocsv
 
 import (
-	"encoding/csv"
 	"errors"
 	"fmt"
-	"io"
 	"reflect"
 )
 
-// Decoder .
-type Decoder interface {
-	getCSVRows() ([][]string, error)
-}
-
-// SimpleDecoder .
-type SimpleDecoder interface {
-	getCSVRow() ([]string, error)
-	getCSVRows() ([][]string, error)
-}
-
-type CSVReader interface {
-	Read() ([]string, error)
-	ReadAll() ([][]string, error)
-}
-
-type csvDecoder struct {
-	CSVReader
-}
-
-func newSimpleDecoderFromReader(r io.Reader) SimpleDecoder {
-	return csvDecoder{getCSVReader(r)}
-}
-
 var (
-	ErrEmptyCSVFile = errors.New("empty csv file given")
 	ErrNoStructTags = errors.New("no csv struct tags found")
 )
-
-// NewSimpleDecoderFromCSVReader creates a SimpleDecoder, which may be passed
-// to the UnmarshalDecoder* family of functions, from a CSV reader. Note that
-// encoding/csv.Reader implements CSVReader, so you can pass one of those
-// directly here.
-func NewSimpleDecoderFromCSVReader(r CSVReader) SimpleDecoder {
-	return csvDecoder{r}
-}
-
-func (c csvDecoder) getCSVRows() ([][]string, error) {
-	return c.ReadAll()
-}
-
-func (c csvDecoder) getCSVRow() ([]string, error) {
-	return c.Read()
-}
 
 func mismatchStructFields(structInfo []fieldInfo, headers []string) []string {
 	missing := make([]string, 0)
@@ -117,263 +74,6 @@ func maybeDoubleHeaderNames(headers []string) error {
 		}
 		headerMap[v] = true
 	}
-	return nil
-}
-
-func readTo(decoder Decoder, out interface{}) error {
-	return readToWithErrorHandler(decoder, nil, out)
-}
-
-func readToWithErrorHandler(decoder Decoder, errHandler ErrorHandler, out interface{}) error {
-	outValue, outType := getConcreteReflectValueAndType(out) // Get the concrete type (not pointer) (Slice<?> or Array<?>)
-	if err := ensureOutType(outType); err != nil {
-		return err
-	}
-	outInnerWasPointer, outInnerType := getConcreteContainerInnerType(outType) // Get the concrete inner type (not pointer) (Container<"?">)
-	if err := ensureOutInnerType(outInnerType); err != nil {
-		return err
-	}
-	csvRows, err := decoder.getCSVRows() // Get the CSV csvRows
-	if err != nil {
-		return err
-	}
-	if len(csvRows) == 0 {
-		return ErrEmptyCSVFile
-	}
-	if err := ensureOutCapacity(&outValue, len(csvRows)); err != nil { // Ensure the container is big enough to hold the CSV content
-		return err
-	}
-	outInnerStructInfo := getStructInfo(outInnerType) // Get the inner struct info to get CSV annotations
-	if len(outInnerStructInfo.Fields) == 0 {
-		return ErrNoStructTags
-	}
-
-	headers := csvRows[0]
-	body := csvRows[1:]
-
-	csvHeadersLabels := make(map[int]*fieldInfo, len(outInnerStructInfo.Fields)) // Used to store the correspondance header <-> position in CSV
-
-	headerCount := map[string]int{}
-	for i, csvColumnHeader := range headers {
-		curHeaderCount := headerCount[csvColumnHeader]
-		if fieldInfo := getCSVFieldPosition(csvColumnHeader, outInnerStructInfo, curHeaderCount); fieldInfo != nil {
-			csvHeadersLabels[i] = fieldInfo
-			if ShouldAlignDuplicateHeadersWithStructFieldOrder {
-				curHeaderCount++
-				headerCount[csvColumnHeader] = curHeaderCount
-			}
-		}
-	}
-
-	if FailIfUnmatchedStructTags {
-		if err := maybeMissingStructFields(outInnerStructInfo.Fields, headers); err != nil {
-			return err
-		}
-	}
-	if FailIfDoubleHeaderNames {
-		if err := maybeDoubleHeaderNames(headers); err != nil {
-			return err
-		}
-	}
-
-	var withFieldsOK bool
-	var fieldTypeUnmarshallerWithKeys TypeUnmarshalCSVWithFields
-
-	for i, csvRow := range body {
-		objectIface := reflect.New(outValue.Index(i).Type()).Interface()
-		outInner := createNewOutInner(outInnerWasPointer, outInnerType)
-		for j, csvColumnContent := range csvRow {
-			if fieldInfo, ok := csvHeadersLabels[j]; ok { // Position found accordingly to header name
-
-				if outInner.CanInterface() {
-					fieldTypeUnmarshallerWithKeys, withFieldsOK = objectIface.(TypeUnmarshalCSVWithFields)
-					if withFieldsOK {
-						if err := fieldTypeUnmarshallerWithKeys.UnmarshalCSVWithFields(fieldInfo.getFirstKey(), csvColumnContent); err != nil {
-							parseError := csv.ParseError{
-								Line:   i + 2, //add 2 to account for the header & 0-indexing of arrays
-								Column: j + 1,
-								Err:    err,
-							}
-							return &parseError
-						}
-						continue
-					}
-				}
-				value := csvColumnContent
-				if value == "" {
-					value = fieldInfo.defaultValue
-				}
-				if err := setInnerField(&outInner, outInnerWasPointer, fieldInfo.IndexChain, value, fieldInfo.omitEmpty); err != nil { // Set field of struct
-					parseError := csv.ParseError{
-						Line:   i + 2, //add 2 to account for the header & 0-indexing of arrays
-						Column: j + 1,
-						Err:    err,
-					}
-					if errHandler == nil || !errHandler(&parseError) {
-						return &parseError
-					}
-				}
-			}
-		}
-
-		if withFieldsOK {
-			reflectedObject := reflect.ValueOf(objectIface)
-			outInner = reflectedObject.Elem()
-		}
-
-		outValue.Index(i).Set(outInner)
-	}
-	return nil
-}
-
-func readEach(decoder SimpleDecoder, c interface{}) error {
-	outValue, outType := getConcreteReflectValueAndType(c) // Get the concrete type (not pointer)
-	if outType.Kind() != reflect.Chan {
-		return fmt.Errorf("cannot use %v with type %s, only channel supported", c, outType)
-	}
-	defer outValue.Close()
-
-	headers, err := decoder.getCSVRow()
-	if err != nil {
-		return err
-	}
-
-	outInnerWasPointer, outInnerType := getConcreteContainerInnerType(outType) // Get the concrete inner type (not pointer) (Container<"?">)
-	if err := ensureOutInnerType(outInnerType); err != nil {
-		return err
-	}
-	outInnerStructInfo := getStructInfo(outInnerType) // Get the inner struct info to get CSV annotations
-	if len(outInnerStructInfo.Fields) == 0 {
-		return ErrNoStructTags
-	}
-	csvHeadersLabels := make(map[int]*fieldInfo, len(outInnerStructInfo.Fields)) // Used to store the correspondance header <-> position in CSV
-	headerCount := map[string]int{}
-	for i, csvColumnHeader := range headers {
-		curHeaderCount := headerCount[csvColumnHeader]
-		if fieldInfo := getCSVFieldPosition(csvColumnHeader, outInnerStructInfo, curHeaderCount); fieldInfo != nil {
-			csvHeadersLabels[i] = fieldInfo
-			if ShouldAlignDuplicateHeadersWithStructFieldOrder {
-				curHeaderCount++
-				headerCount[csvColumnHeader] = curHeaderCount
-			}
-		}
-	}
-	if err := maybeMissingStructFields(outInnerStructInfo.Fields, headers); err != nil {
-		if FailIfUnmatchedStructTags {
-			return err
-		}
-	}
-	if FailIfDoubleHeaderNames {
-		if err := maybeDoubleHeaderNames(headers); err != nil {
-			return err
-		}
-	}
-	i := 0
-	for {
-		line, err := decoder.getCSVRow()
-		if err == io.EOF {
-			break
-		} else if err != nil {
-			return err
-		}
-		outInner := createNewOutInner(outInnerWasPointer, outInnerType)
-		for j, csvColumnContent := range line {
-			if fieldInfo, ok := csvHeadersLabels[j]; ok { // Position found accordingly to header name
-				if err := setInnerField(&outInner, outInnerWasPointer, fieldInfo.IndexChain, csvColumnContent, fieldInfo.omitEmpty); err != nil { // Set field of struct
-					return &csv.ParseError{
-						Line:   i + 2, //add 2 to account for the header & 0-indexing of arrays
-						Column: j + 1,
-						Err:    err,
-					}
-				}
-			}
-		}
-		outValue.Send(outInner)
-		i++
-	}
-	return nil
-}
-
-func readEachWithoutHeaders(decoder SimpleDecoder, c interface{}) error {
-	outValue, outType := getConcreteReflectValueAndType(c) // Get the concrete type (not pointer) (Slice<?> or Array<?>)
-	if err := ensureOutType(outType); err != nil {
-		return err
-	}
-	defer outValue.Close()
-
-	outInnerWasPointer, outInnerType := getConcreteContainerInnerType(outType) // Get the concrete inner type (not pointer) (Container<"?">)
-	if err := ensureOutInnerType(outInnerType); err != nil {
-		return err
-	}
-	outInnerStructInfo := getStructInfo(outInnerType) // Get the inner struct info to get CSV annotations
-	if len(outInnerStructInfo.Fields) == 0 {
-		return ErrNoStructTags
-	}
-
-	i := 0
-	for {
-		line, err := decoder.getCSVRow()
-		if err == io.EOF {
-			break
-		} else if err != nil {
-			return err
-		}
-		outInner := createNewOutInner(outInnerWasPointer, outInnerType)
-		for j, csvColumnContent := range line {
-			fieldInfo := outInnerStructInfo.Fields[j]
-			if err := setInnerField(&outInner, outInnerWasPointer, fieldInfo.IndexChain, csvColumnContent, fieldInfo.omitEmpty); err != nil { // Set field of struct
-				return &csv.ParseError{
-					Line:   i + 2, //add 2 to account for the header & 0-indexing of arrays
-					Column: j + 1,
-					Err:    err,
-				}
-			}
-		}
-		outValue.Send(outInner)
-		i++
-	}
-	return nil
-}
-
-func readToWithoutHeaders(decoder Decoder, out interface{}) error {
-	outValue, outType := getConcreteReflectValueAndType(out) // Get the concrete type (not pointer) (Slice<?> or Array<?>)
-	if err := ensureOutType(outType); err != nil {
-		return err
-	}
-	outInnerWasPointer, outInnerType := getConcreteContainerInnerType(outType) // Get the concrete inner type (not pointer) (Container<"?">)
-	if err := ensureOutInnerType(outInnerType); err != nil {
-		return err
-	}
-	csvRows, err := decoder.getCSVRows() // Get the CSV csvRows
-	if err != nil {
-		return err
-	}
-	if len(csvRows) == 0 {
-		return ErrEmptyCSVFile
-	}
-	if err := ensureOutCapacity(&outValue, len(csvRows)+1); err != nil { // Ensure the container is big enough to hold the CSV content
-		return err
-	}
-	outInnerStructInfo := getStructInfo(outInnerType) // Get the inner struct info to get CSV annotations
-	if len(outInnerStructInfo.Fields) == 0 {
-		return ErrNoStructTags
-	}
-
-	for i, csvRow := range csvRows {
-		outInner := createNewOutInner(outInnerWasPointer, outInnerType)
-		for j, csvColumnContent := range csvRow {
-			fieldInfo := outInnerStructInfo.Fields[j]
-			if err := setInnerField(&outInner, outInnerWasPointer, fieldInfo.IndexChain, csvColumnContent, fieldInfo.omitEmpty); err != nil { // Set field of struct
-				return &csv.ParseError{
-					Line:   i + 1,
-					Column: j + 1,
-					Err:    err,
-				}
-			}
-		}
-		outValue.Index(i).Set(outInner)
-	}
-
 	return nil
 }
